[libc++] Refactor the tests for [iterator.range]
The tests were a bit of a mess -- the testing coverage wasn't bad but
it was extremely difficult to see what was being tested and where.
I split up the tests to make them easier to audit for completeness and
did such an audit, adding a few missing tests (e.g. the conditional
noexcept-ness of std::cbegin and std::cend). I also audited the synopsis
and adjusted it where it needed to be adjusted.

This patch is in preparation of fixing #67471.
